fix(health): attribute Fitbit samples, dedup sessions, compact heart rate
Three defects found while auditing the health data that has accumulated
since the Fitbit Air started publishing on 2026-07-25. They ship together
because they share the `source` column's meaning.

source attribution
  Fitbit-native points carry no `dataSource.application.packageName` — only
  `platform` ("FITBIT") + `recordingMethod` — so the parser's package-name-only
  read dropped 174,561 rows (98% of them heart rate) into one opaque "unknown"
  bucket. `sampleSource()` now falls back to the platform, and migration 0035
  relabels the stored rows. The two MUST ship together: `source` is part of the
  sample identity, so leaving them as 'unknown' would make the next sync
  re-insert all of them under the new key. Scope is provably safe — the
  on-device importer falls back to 'healthconnect', so nothing else writes
  'unknown'. Also stops non-string attribution being coerced into the column.

session dedup
  /api/fitbit/summary deduped exercise by exact start but had no equivalent for
  sleep, so a night stored by both the Fitbit and com.withings.wiscale2 rendered
  twice at different stage granularity — the 10-night list showed 7 nights. The
  Withings app is an aggregator, not a sensor: it re-publishes what it reads from
  Health Connect (identical DEEP/REM minutes; 36 of 39 of its workouts start in
  the same second as another source's; its history predates the Fitbit). New
  modules/health/sessions.ts keys on the start second — which also catches the
  same workout republished 389ms apart, inflating four days of exercise minutes
  (07-20 read 78 min against an actual 39) — and prefers the measuring platform.

heart-rate compaction
  Heart rate was 98.7% of health_samples (287 MB of 291 MB) and projected to
  ~3.2 GB/yr at the Fitbit's ~2.3s cadence. A daily job now compacts settled
  ranges into per-minute buckets carrying { min, max, n }, verified on real data
  at 37,616 rows -> 1,414 (26.6x) with the daily mean identical to nine decimals
  and min/max/count bit-identical. Verbatim payloads were already archived in
  health_raw_pages (the same 786,336 points compress to 11 MB there), so nothing
  becomes unrecoverable. Compacting closed ranges rather than bucketing at ingest
  keeps it idempotent: the contributing raw rows are deleted in the same
  transaction, so the merge cannot double-count. recomputeDailySummaries and the
  summary route's live-aggregation fallback both read the stored bounds, so the
  per-day range bar stays honest.

data-usage counters
  data_usage_cache.estimated_bytes was int4 and the insights per-category
  SUM(...)::int would have thrown "integer out of range" — location was at 75% of
  the 2 GB ceiling, growing ~7.8 MB/day. Widened both (migration 0036).

+++ src/modules/insights/service.ts
@@ -1069,8 +1069,12 @@ export class InsightsService {
     const rows = await db
       .select({
         category: dataUsageCache.category,
-        bytes: sql<number>`SUM(${dataUsageCache.estimatedBytes})::int`,
-        rows: sql<number>`SUM(${dataUsageCache.rowCount})::int`,
+        // ::bigint, not ::int — a category's summed footprint passes int4's 2 GB
+        // ceiling (location was at 1.5 GB here), which would throw "integer out of
+        // range" and take out the whole data-usage section. Read back through
+        // Number() below, since pg returns bigint as a string.
+        bytes: sql<number>`SUM(${dataUsageCache.estimatedBytes})::bigint`,
+        rows: sql<number>`SUM(${dataUsageCache.rowCount})::bigint`,
       })
       .from(dataUsageCache)
       .where(eq(dataUsageCache.userId, userId))
